--- a/source/numerics/CMakeLists.txt
+++ b/source/numerics/CMakeLists.txt
@@ -50,6 +50,7 @@ SET(_src
   vector_tools_project.cc
   vector_tools_project_inst2.cc
   vector_tools_project_inst3.cc
+  vector_tools_project_hp.cc
   vector_tools_project_qp.cc
   vector_tools_project_qpmf.cc
   vector_tools_rhs.cc
@@ -80,6 +81,7 @@ SET(_inst
   vector_tools_point_value.inst.in
   vector_tools_point_gradient.inst.in
   vector_tools_project.inst.in
+  vector_tools_project_hp.inst.in
   vector_tools_project_qp.inst.in
   vector_tools_project_qpmf.inst.in
   vector_tools_rhs.inst.in