--- a/source/numerics/vector_tools_project_hp.inst.in
+++ b/source/numerics/vector_tools_project_hp.inst.in
@@ -1,6 +1,6 @@
 // ---------------------------------------------------------------------
 //
-// Copyright (C) 1998 - 2014 by the deal.II authors
+// Copyright (C) 2017 by the deal.II authors
 //
 // This file is part of the deal.II library.
 //
@@ -20,26 +20,26 @@ for (VEC : REAL_SERIAL_VECTORS ; deal_II_dimension : DIMENSIONS; deal_II_space_d
     namespace VectorTools \{
 
     template
-    void project<deal_II_dimension, VEC, deal_II_space_dimension>
-    (const Mapping<deal_II_dimension,deal_II_space_dimension>      &,
-     const DoFHandler<deal_II_dimension,deal_II_space_dimension>   &,
+    void project
+    (const hp::MappingCollection<deal_II_dimension,deal_II_space_dimension>      &,
+     const hp::DoFHandler<deal_II_dimension,deal_II_space_dimension>   &,
      const ConstraintMatrix                &,
-     const Quadrature<deal_II_dimension>   &,
+     const hp::QCollection<deal_II_dimension>   &,
      const Function<deal_II_space_dimension,VEC::value_type>     &,
      VEC                                   &,
      const bool,
-     const Quadrature<deal_II_dimension-1> &,
+     const hp::QCollection<deal_II_dimension-1> &,
      const bool);
 
     template
-    void project<deal_II_dimension, VEC, deal_II_space_dimension>
-    (const DoFHandler<deal_II_dimension,deal_II_space_dimension>   &,
+    void project
+    (const hp::DoFHandler<deal_II_dimension,deal_II_space_dimension>   &,
      const ConstraintMatrix                &,
-     const Quadrature<deal_II_dimension>   &,
+     const hp::QCollection<deal_II_dimension>   &,
      const Function<deal_II_space_dimension,VEC::value_type>     &,
      VEC                                   &,
      const bool,
-     const Quadrature<deal_II_dimension-1> &,
+     const hp::QCollection<deal_II_dimension-1> &,
      const bool);
     \}
 #endif