I spent a little time looking into ways to help #3825 and came across one that is pretty easy to do.

This new file takes about 17 seconds to compile and improves the performance of the non-hp (i.e., the matrix free) projections a little bit.

before splitting out the hp files:
```
                                    File name Memory, MB    Time, s
      source/numerics/vector_tools_project.cc       2860         81
source/numerics/vector_tools_project_inst2.cc       2970        100
source/numerics/vector_tools_project_inst3.cc       3956        114
```

After:
```
                                    File name Memory, MB    Time, s
      source/numerics/vector_tools_project.cc       2775         79
source/numerics/vector_tools_project_inst2.cc       2882         92
source/numerics/vector_tools_project_inst3.cc       3948        102
```
These numbers come from setting the C++ compiler as a shell script containing
```sh
/usr/bin/time -v g++ ${1+"$@"}
```
and then parsing the output.