@drwells 2017-01-22
I wrote a second patch that moves all of the codimension 1 and 2 instantiations into a second file:

```
                                    File name Memory, MB    Time, s
source/numerics/vector_tools_project_codim.cc       1843         27
      source/numerics/vector_tools_project.cc       2773         77
source/numerics/vector_tools_project_inst2.cc       2832         81
source/numerics/vector_tools_project_inst3.cc       2822         85
```

I will put up a PR for that once this one is merged since I changed adjacent lines in the `CMakeLists.txt` file.