@bangerth 2017-01-22
By the way, if you used `ninja` as your build tool as a lot of us do these days, you don't even have to wrap the compiler, but can get run times from the `.ninja_log` file that's in the build directory. The first and second columns are the start and end time of the compile job in milliseconds. (ninja uses this information to sort the order of targets if you build stuff again later, but it's also useful in itself.)