@drwells 2017-01-22
That is a useful point. Is there any easy way to also get an estimate of memory consumption from ninja?